// ankomstregistreringExpo/src/models/LoadedBokning.ts
 export default interface LoadedBokning {
    LateTxt: string;
    SchSNr: number;
    EmpPDsc: string;
    SchPDsc: string;
    DatTimeEnd: string;
    WelcomeTxt: string;
    ToLate: boolean;
    LateText: string;
    DatTimeStart: number;
    Stat: number;
    DatSch: number;
    TimeStart: number;
    TimeEnd: number;
    TabYAATab: number;
    TabYAANr: number;
  }